Adds TCK component
- Refactors InMemory code and TCK
- Refactors Aas Discovery component
- Refactors AAS Discovery Service Backend InMemory
- Refactors AAS Discovery Service Backend InMemory
- Refactors AAS Discovery Service Backend MongoDB
- Refactors AAS Discovery Service core
- Refactors AAS Discovery Service http
- Updates Readme
- Refactors TCK name
