+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice-backend-inmemory/src/test/java/org/eclipse/digitaltwin/basyx/aasdiscoveryservice/backend/inmemory/TestInMemoryAasDiscoveryService.java
@@ -23,41 +23,23 @@
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
 
-package org.eclipse.digitaltwin.basyx.aasdiscoveryservice.http.testconfig;
 
-import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.backend.mongodb.MongoDBAasDiscoveryService;
+package org.eclipse.digitaltwin.basyx.aasdiscoveryservice.backend.inmemory;
+
 import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryService;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.core.MongoTemplate;
-
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
+import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryServiceSuite;
 
 /**
- * Configuration for tests
+ * Tests the {@link InMemoryAasDiscoveryService}
  * 
- * @author danish
+ * @author zhangzai
  *
  */
-@Configuration
-public class DummyConfig {
-
-	private final String COLLECTION = "discoveryServiceHTTPTestCollection";
-
-	@Bean
-	@ConditionalOnMissingBean
-	public AasDiscoveryService createAasDiscoveryService() {
-		return new MongoDBAasDiscoveryService(createTemplate(), COLLECTION);
-	}
-
-	private MongoTemplate createTemplate() {
-		String connectionURL = "mongodb://mongoAdmin:mongoPassword@localhost:27017/";
-
-		MongoClient client = MongoClients.create(connectionURL);
-
-		return new MongoTemplate(client, "BaSyxTestDb");
+public class TestInMemoryAasDiscoveryService extends AasDiscoveryServiceSuite {
+	
+	@Override
+	protected AasDiscoveryService getAasDiscoveryService() {
+		return new InMemoryAasDiscoveryService();
 	}
 
 }
