+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice-backend-mongodb/src/main/java/org/eclipse/digitaltwin/basyx/aasdiscoveryservice/backend/mongodb/MongoDBAasDiscoveryServiceFactory.java
@@ -1,0 +1,68 @@
+/*******************************************************************************
+ * Copyright (C) 2023 the Eclipse BaSyx Authors
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * 
+ * SPDX-License-Identifier: MIT
+ ******************************************************************************/
+
+package org.eclipse.digitaltwin.basyx.aasdiscoveryservice.backend.mongodb;
+
+import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryService;
+import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryServiceFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
+import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.stereotype.Component;
+
+/**
+ * {@link AasDiscoveryService} factory returning a MongoDB backend
+ * 
+ * @author danish
+ */
+@Component
+@ConditionalOnExpression("'${basyx.backend}'.equals('MongoDB')")
+public class MongoDBAasDiscoveryServiceFactory implements AasDiscoveryServiceFactory {
+
+	private MongoTemplate mongoTemplate;
+	private String collectionName;
+	private String aasDiscoveryServiceName;
+
+	@Autowired(required = false)
+	public MongoDBAasDiscoveryServiceFactory(MongoTemplate mongoTemplate,
+			@Value("${basyx.aasdiscoveryservice.mongodb.collectionName:aasdiscovery-service}") String collectionName) {
+		this.mongoTemplate = mongoTemplate;
+		this.collectionName = collectionName;
+	}
+
+	@Autowired(required = false)
+	public MongoDBAasDiscoveryServiceFactory(MongoTemplate mongoTemplate,
+			@Value("${basyx.aasdiscoveryservice.mongodb.collectionName:aasdiscovery-service}") String collectionName,
+			@Value("${basyx.aasdiscserv.name:aas-discovery-service}") String aasDiscoveryServiceName) {
+		this(mongoTemplate, collectionName);
+		this.aasDiscoveryServiceName = aasDiscoveryServiceName;
+	}
+
+	@Override
+	public AasDiscoveryService create() {
+		return new MongoDBAasDiscoveryService(mongoTemplate, collectionName, aasDiscoveryServiceName);
+	}
+}
