+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice-core/src/main/java/org/eclipse/digitaltwin/basyx/aasdiscoveryservice/core/feature/DecoratedAasDiscoveryServiceFactory.java
@@ -23,41 +23,28 @@
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
 
-package org.eclipse.digitaltwin.basyx.aasdiscoveryservice.http.testconfig;
+package org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.feature;
 
-import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.backend.mongodb.MongoDBAasDiscoveryService;
+import java.util.List;
+
 import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryService;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.core.MongoTemplate;
-
-import com.mongodb.client.MongoClient;
-import com.mongodb.client.MongoClients;
+import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryServiceFactory;
+import org.eclipse.digitaltwin.basyx.core.DecoratedFactory;
 
 /**
- * Configuration for tests
+ * Factory for {@link AasDiscoveryService} decoration
  * 
- * @author danish
+ * @author schnicke, danish, zhangzai
  *
  */
-@Configuration
-public class DummyConfig {
+public class DecoratedAasDiscoveryServiceFactory extends DecoratedFactory<AasDiscoveryServiceFactory, AasDiscoveryServiceFeature> implements AasDiscoveryServiceFactory {
 
-	private final String COLLECTION = "discoveryServiceHTTPTestCollection";
-
-	@Bean
-	@ConditionalOnMissingBean
-	public AasDiscoveryService createAasDiscoveryService() {
-		return new MongoDBAasDiscoveryService(createTemplate(), COLLECTION);
+	public DecoratedAasDiscoveryServiceFactory(AasDiscoveryServiceFactory toDecorate, List<AasDiscoveryServiceFeature> features) {
+		super(toDecorate, features);
 	}
 
-	private MongoTemplate createTemplate() {
-		String connectionURL = "mongodb://mongoAdmin:mongoPassword@localhost:27017/";
-
-		MongoClient client = MongoClients.create(connectionURL);
-
-		return new MongoTemplate(client, "BaSyxTestDb");
+	@Override
+	public AasDiscoveryService create() {
+		return getDecorated().create();
 	}
-
 }
