+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice-http/src/main/java/org/eclipse/digitaltwin/basyx/aasdiscoveryservice/http/documentation/AasDiscoveryServiceApiDocumentationConfiguration.java
@@ -25,13 +25,14 @@
 
 package org.eclipse.digitaltwin.basyx.aasdiscoveryservice.http.documentation;
 
+import org.eclipse.digitaltwin.basyx.aasdiscoveryservice.core.AasDiscoveryService;
 import org.eclipse.digitaltwin.basyx.http.documentation.RepositoryApiDocumentationConfiguration;
 import org.springframework.context.annotation.Configuration;
 
 import io.swagger.v3.oas.models.info.Info;
 
 /**
- * API documentation configuration for {@link ConceptDescriptionRepository}
+ * API documentation configuration for {@link AasDiscoveryService}
  * 
  * @author danish
  *
