--- basyx.aasdiscoveryservice/basyx.aasdiscoveryservice-http/src/test/resources/AssetIds.json
@@ -1,4 +1,0 @@
-[
-	"DummyAsset_1_Value",
-	"DummyAsset_3_Value"
-]
