+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice.component/pom.xml
@@ -12,7 +12,7 @@
 
 	<properties>
 		<docker.image.name>aasdiscovery-service</docker.image.name>
-		<docker.container.waitForEndpoint>http://localhost:${docker.host.port}/aasdiscovery</docker.container.waitForEndpoint>
+		<docker.container.waitForEndpoint>http://localhost:${docker.host.port}/lookup/shells</docker.container.waitForEndpoint>
 	</properties>
 
 	<dependencies>
