+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice.component/src/main/resources/application.properties
@@ -1,10 +1,10 @@
 server.port=8081
-spring.application.name=Concept Description Repository
-basyx.cdrepo.name = cd-repo
+spring.application.name=AAS Discovery Service
+basyx.aasdiscserv.name=aas-discovery-service
 
-basyx.backend = InMemory
+basyx.backend=InMemory
 
-#basyx.backend = MongoDB
+#basyx.backend=MongoDB
 #spring.data.mongodb.host=127.0.0.1
 ##or spring.data.mongodb.host=127.0.0.1
 #spring.data.mongodb.port=27017
@@ -12,6 +12,7 @@ basyx.backend = InMemory
 #spring.data.mongodb.authentication-database=admin
 #spring.data.mongodb.username=mongoAdmin
 #spring.data.mongodb.password=mongoPassword
+#basyx.aasdiscoveryservice.mongodb.collectionName=aasdiscovery-service
 
 #alternative
 #spring.data.mongodb.uri=mongodb://mongoAdmin:mongoPassword@localhost:27017/?authMechanism=DEFAULT
