+++ basyx.aasdiscoveryservice/basyx.aasdiscoveryservice.component/src/main/resources/banner.txt
@@ -5,5 +5,5 @@
  | |_) || (_| | ____) || |_| | >  < 
  |____/  \__,_||_____/  \__, |/_/\_\
 ======================== __/ |======      
-Concept-Desc Repository |___/                  
+AAS Discovery Service   |___/                  
 2.0.0-PREVIEW
