--- basyx.aasdiscoveryservice/basyx.aasdiscoveryservice.component/src/test/java/org/eclipse/digitaltwin/basyx/conceptdescriptionrepository/component/ConceptDescriptionRepositoryIT.java
@@ -1,12 +1,0 @@
-package org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.component;
-
-import org.eclipse.digitaltwin.basyx.aasrepository.tck.ConceptDescriptionRepositoryTestDefinedURL;
-
-/**
- * Integration test using the default "url" endpoint from the TCK
- * 
- * @author espen
- *
- */
-public class ConceptDescriptionRepositoryIT extends ConceptDescriptionRepositoryTestDefinedURL {
-}
