+++ basyx.aasdiscoveryservice/pom.xml
@@ -16,8 +16,8 @@
 		<module>basyx.aasdiscoveryservice-core</module>
 		<module>basyx.aasdiscoveryservice-backend-mongodb</module>
 		<module>basyx.aasdiscoveryservice-http</module>
-		<!-- <module>basyx.aasdiscoveryservice-backend-inmemory</module>
-		
-		<module>basyx.aasdiscoveryservice.component</module> -->
+		<module>basyx.aasdiscoveryservice-backend-inmemory</module>
+		<module>basyx.aasdiscoveryservice-tck</module>
+		<module>basyx.aasdiscoveryservice.component</module>
 	</modules>
 </project>
